fix bug in reader
